C++: Track changes to the value for reference parameters in SSA
This is work in preparation for supporting reference parameters in
range analysis.

With local reference variables, the variable is an alias for an existing
memory location within the same scope, which makes tracking changes to
the value difficult. For reference parameters, the aliasing occurs
outside of the scope of the function, which makes it less likely that
there will be another alias in scope. For this reason, I think it's safe
to track the changes to values for reference parameters.