Java: XXE - flag results with at least one unsafe SAXParser
The existing logic for identifying `isSafe()` SAXParsers specified:

    exists(SafeSAXParserFlowConfig sp |
      sp.hasFlowToExpr(this.getQualifier())
    )

Which specifies that a SAXParserParse call is safe if there exists at
least one SafeSAXParser that flows to the qualifier of this call. This
approach can cause false negatives where either multiple SAXParsers can
flow to single SAXParser.parse call, or where a SAXParser can be created
from multiple SAXParserFactory instances, and there is at least one safe
and one unsafe instance.

This commit duplicates the SAXParser logic to identify "unsafe"
SAXParsers as well as "safe" SAXParsers, then updates the `isSafe()`
predicate to say a SAXParserParse call is only safe if:
 - There are no unsafe parsers that flow to this call.
 - There is at least one safe parser that flows to this call.

This is a straightforward way to add this logic, but it does add more
duplication to an already quite boilerplate heavy library. I suspect
there could be some advantages to templating the whole approach here to
reduce duplication.