Java: Make smap remapping visible in Locations
If an SMAP line mapping file is included in the database, we remap
hasLocationInfo on Top if we find a mapping. However, we do not change
the result of .getLocation(). This is problematic for queries which
include data flow nodes in the select clause, because DataFlow calls
getLocation() on the underlying node to return location info, rather
than calling hasLocationInfo.

I have modified the Location itself to report the remap line numbers so
that we get a consitent view of location for that element. We do not
necessarily have Location entries for the targets of the remapping, so
we cannot just report the Location in the original file.